Fixes #2: Completed extract.py
Added logic to extract text from Wikipedia sections related to education or early life. Handles missing pages and improves robustness of section detection. Additionally, I left the openpyxl excel iteration to be higher level, once the NLP pipeline is finished we can use this in tandem with the NLP pipeline in a simple for-loop.